guardian_daemon/pyproject.toml
[project]
name = "guardian_daemon"
version = "0.1.0"
description = "Multi-device parental control system - Daemon"
authors = [{ name = "Phillip Stockmann", email = "ZyanKLee@users.noreply.github.com" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbus-next>=0.2.3",
    "pyyaml>=6.0.2",
    "structlog>=25.4.0",
]

[dependency-groups]
dev = [
    "pip>=25.2",
    "types-pyyaml>=6.0.12.20250915",
]


